convert pinned listing
